fix: separate project detail tab header
